fix: uso do 'email' nas permissões de vez 'numeroMatricula'
